testing pricing standard

+++ crates/brontes-database/src/libmdbx/initialize.rs
@@ -721,7 +721,7 @@ mod tests {
                 let new_pairs = new_mapping.get(&old_exch).unwrap();
                 println!("NEW PAIRS: {}", new_pairs.len());
                 println!("OLD PAIRS: {}\n", old_pairs.len());
-                assert!(new_pairs.len() >= old_pairs.len() * 0.5);
+                assert!(new_pairs.len() as f64 >= old_pairs.len() as f64 * 0.5);
             })
         });
